Sync with upstream v275

+++ lib/language_pack/ruby.rb
@@ -701,19 +701,35 @@ BUNDLE
       end
 
       if bundler.windows_gemfile_lock?
-        log("bundle", "has_windows_gemfile_lock")
+        if bundler.supports_multiple_platforms?
+          puts "Windows platform detected, preserving `Gemfile.lock` (#{bundler.version} >= 2.2)"
+        else
+            File.unlink("Gemfile.lock")
+            ENV.delete("BUNDLE_DEPLOYMENT")
 
-        File.unlink("Gemfile.lock")
-        ENV.delete("BUNDLE_DEPLOYMENT")
+            warn(<<~WARNING, inline: true)
+              Removing `Gemfile.lock` because it was generated on Windows.
 
-        warn(<<~WARNING, inline: true)
-          Removing `Gemfile.lock` because it was generated on Windows.
-          Bundler will do a full resolve so native gems are handled properly.
-          This may result in unexpected gem versions being used in your app.
-          In rare occasions Bundler may not be able to resolve your dependencies at all.
+              Bundler will do a full resolve so native gems are handled properly.
+              This may result in unexpected gem versions being used in your app.
+              In rare occasions Bundler may not be able to resolve your dependencies at all.
 
-          https://doc.scalingo.com/languages/ruby/bundle-gemfile-on-windows
-        WARNING
+              Upgrade to bundler version 2.2 or higher to skip this behavior and use built in support
+              for multiple platforms. Running these commands below (after the `>`) in a command prompt
+              should resolve this warning:
+
+                > gem install bundler
+                > bundle update --bundler
+                > bundle lock --add-platform ruby
+                > bundle lock --add-platform x86_64-linux
+                > bundle install
+                > git add Gemfile.lock
+                > git commit -m "Upgrade bundler"
+
+              For more information see:
+                https://doc.scalingo.com/languages/ruby/bundle-gemfile-on-windows
+            WARNING
+        end
       end
 
       bundle_command = String.new("")
